# Use ringbuffer for windowing functions