--- a/ringbuffer/ringbuffer.go
+++ b/ringbuffer/ringbuffer.go
@@ -3,6 +3,8 @@
 
 package ringbuffer
 
+import "math"
+
 type Sample[T any] struct {
 	T int64
 	V T
@@ -23,7 +25,12 @@ func (r *RingBuffer[T]) Len() int {
 	return len(r.items)
 }
 
+// MaxT returns the maximum timestamp of the ring buffer.
+// If the ring buffer is empty, it returns math.MinInt64.
 func (r *RingBuffer[T]) MaxT() int64 {
+	if len(r.items) == 0 {
+		return math.MinInt64
+	}
 	return r.items[len(r.items)-1].T
 }
 
@@ -35,6 +42,10 @@ func (r *RingBuffer[T]) ReadIntoNext(f func(*Sample[T])) {
 		r.items = append(r.items, Sample[T]{})
 	}
 	f(&r.items[n])
+}
+
+func (r *RingBuffer[T]) ReadIntoLast(f func(*Sample[T])) {
+	f(&r.items[len(r.items)-1])
 }
 
 func (r *RingBuffer[T]) Push(t int64, v T) {
@@ -54,6 +65,26 @@ func (r *RingBuffer[T]) DropBefore(ts int64) {
 	}
 	var drop int
 	for drop = 0; drop < len(r.items) && r.items[drop].T < ts; drop++ {
+	}
+	keep := len(r.items) - drop
+
+	r.tail = resize(r.tail, drop)
+	copy(r.tail, r.items[:drop])
+	copy(r.items, r.items[drop:])
+	copy(r.items[keep:], r.tail)
+	r.items = r.items[:keep]
+}
+
+func (r *RingBuffer[T]) DropBeforeWithExtLookback(ts int64, extMint int64) {
+	if len(r.items) == 0 || r.items[len(r.items)-1].T < ts {
+		r.items = r.items[:0]
+		return
+	}
+	var drop int
+	for drop = 0; drop < len(r.items) && r.items[drop].T < ts; drop++ {
+	}
+	if drop > 0 && r.items[drop-1].T >= extMint {
+		drop--
 	}
 	keep := len(r.items) - drop